% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circlecorR-package.R
\docType{package}
\name{circlecorR-package}
\alias{circlecorR}
\alias{circlecorR-package}
\title{circlecorR: Circular Correlation Wheel Plots}
\description{
Draws MNE-style circular "correlation wheel" plots natively in R. The main
entry point is \code{\link[=corr_wheel]{corr_wheel()}}; \code{\link[=compute_correlations]{compute_correlations()}} builds the r/p input
from raw data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kriz98.github.io/circlecorR/}
  \item \url{https://github.com/kriz98/circlecorR}
  \item Report bugs at \url{https://github.com/kriz98/circlecorR/issues}
}

}
\author{
\strong{Maintainer}: Chris Varghese \email{chris.varghese@auckland.ac.nz}

Authors:
\itemize{
  \item Chris Varghese \email{chris.varghese@auckland.ac.nz}
  \item Armen A. Gharibans
}

}
\keyword{internal}
