% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemes.R
\name{corr_wheel_schemes}
\alias{corr_wheel_schemes}
\title{List built-in colour schemes}
\usage{
corr_wheel_schemes()
}
\value{
A character vector of scheme names.
}
\description{
Returns the names of the colour schemes bundled with the package, for use
as the \code{scheme} argument of \code{\link[=corr_wheel]{corr_wheel()}}.
}
\examples{
corr_wheel_schemes()
}
\seealso{
\code{\link[=corr_wheel_scheme]{corr_wheel_scheme()}}, \code{\link[=corr_wheel]{corr_wheel()}}
}
