% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gastro_cor}
\alias{gastro_cor}
\title{Correlation matrices for the synthetic gastric-symptom dataset}
\format{
A list of class \code{"circlecor"} with elements \code{r} (16x16 correlation
matrix) and \code{p} (16x16 p-value matrix).
}
\usage{
gastro_cor
}
\description{
A \code{"circlecor"} object (Pearson, Hochberg-adjusted) computed from
\link{gastro_symptoms} with \code{\link[=compute_correlations]{compute_correlations()}}. Ready to pass to
\code{\link[=corr_wheel]{corr_wheel()}}.
}
\seealso{
\link{gastro_symptoms}, \code{\link[=corr_wheel]{corr_wheel()}}
}
\keyword{datasets}
